Rename oauthClient to webhookProvider and move it to shared
